OpenFaaS/Functions/Users/delete-user/requirements.txt
Flask==2.3.2
psycopg2-binary==2.9.6
pyjwt==2.8.0